Reg.With.Mohs.D
Our goal in this project is to use regression to predict the Mohs hardness of a mineral given its properties.
Mohs hardness is a system of scales used to compare the hardness of minerals. This scale includes values ​​from 1 to 10 and is based on the mineral's ability to scratch other minerals. The Mohs hardness scale was developed by the German mineralogist Friedrich Mohs in 1812.
For example, talc is a softer mineral and scratches easily on the diamond side.